Program Information
The program is designed to perform an inventory of basic GIS data such as .lyr, .mdb, .gdb and .shp files. While it was designed for GIS inventory of a specific organization, the code can be easily modified to accommodate to the needs of various organizations. The program outputs several csv files containing inventory data.